重構: 抽離假資料、型別定義與狀態處理邏輯至獨立模組
